+++ src/github.com/influxdb/influxdb/src/parser/frees.c
@@ -46,6 +46,9 @@ free_value_array(value_array *array)
 void
 free_groupby_clause(groupby_clause *g)
 {
+  if (!g)
+    return;
+
   free_value_array(g->elems);
   if (g->fill_function) {
     free_value(g->fill_function);
@@ -96,6 +99,11 @@ free_select_query (select_query *q)
     free_groupby_clause(q->group_by);
   }
 
+  if (q->into_clause) {
+    free_value(q->into_clause->target);
+    free(q->into_clause);
+  }
+
   if (q->from_clause) {
     // free the from clause
     free_from_clause(q->from_clause);
@@ -116,6 +124,12 @@ free_delete_query (delete_query *q)
 }
 
 void
+free_drop_series_query (drop_series_query *q)
+{
+  free_value(q->name);
+}
+
+void
 close_query (query *q)
 {
    if (q->error) {
@@ -125,6 +139,15 @@ close_query (query *q)
   if (q->select_query) {
     free_select_query(q->select_query);
     free(q->select_query);
+  }
+
+  if (q->drop_series_query) {
+    free_drop_series_query(q->drop_series_query);
+    free(q->drop_series_query);
+  }
+
+  if (q->drop_query) {
+    free(q->drop_query);
   }
 
   if (q->delete_query) {